+++ lcg_parallel2.cpp
@@ -5,9 +5,9 @@ LCG_Parallel2 :: LCG_Parallel2(int number_of_processes)
 	num_procs = number_of_processes;	
 }
 
-vector<int> LCG_Parallel2 :: generate_random_numbers(int count, int seed, pair<int, int> range)
+vector<double> LCG_Parallel2 :: generate_random_numbers(int count, int seed, pair<int, int> range)
 {
 	cout<<"Method 2"<<endl;
-	vector<int> tmp;
+	vector<double> tmp;
 	return tmp;
 }
